requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
reportlab==4.0.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
